requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
coloredlogs==15.0.1
comm==0.2.1
dataclasses==0.6
dataclasses-json==0.6.3
debugpy==1.6.7
decorator==5.1.1
diskcache==5.6.3
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.108.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
grpcio==1.60.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.28.0
ipython==8.12.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter_core==5.5.0
langchain==0.0.339
langchain-community==0.0.8
langchain-core==0.1.5
langsmith==0.0.77
llama_cpp_python==0.2.26
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
onnxruntime==1.16.3
overrides==7.4.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.1.0
posthog==3.1.0
prompt-toolkit==3.0.42
protobuf==4.25.1
psutil==5.9.0
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.12
PyPika==0.48.9
python-box==7.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.16.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
stack-data==0.6.2
starlette==0.32.0.post1
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.1
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
typer==0.9.0
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
wheel==0.41.2
yarl==1.9.4
zipp==3.17.0
